refactor: Add order.simplequery package and separate v4 logic
